+++ ChangeLog.md
@@ -1,9 +1,18 @@
 # OPENGL IMGUI
 
+## [0.0.3]  date : 2024-4-24
+[+] 更新ReadMe.md笔记，增加有关uniform相关的内容
+[+] 增加第3个demo - src/shader03 《着色器》
+[+] 增加获取运行时间的接口 glfwGetTime
+[+] 增加IMGUI-Color部分,对颜色进行选择
+[+] 增加顶点属性,使用glVertexAttribPointer进行映射
+[+] 通过全局变量,实现动态数据的映射(glBindBuffer)
+
 ## [0.0.2]  date : 2024-4-22
 [*] 更新ReadMe.md笔记，补充有关缓冲对象相关内容
 [+] 编写第2个demo- src/triangle02 《你好，三角形》
 [+] 增加IMGUI的依赖，在项目中增加IMGUI
+[+] 增加代码区域
 
 
 ## [0.0.1]  date : 2024-4-18